Compile mess.exs instead of evaluating it.
